Add files via upload
Takes an input dictionary and plays a game of Hangman. Does not pick a word until the end of the game, when either there is only one possible word from the dictionary or the player runs out of guesses. Accomplishes this with a map of Strings to Set<String> and taking the largest possible set of strings.